Changed reader implementation to use ParquetReader
* Previous implementation kept as TablesawPaquetFileReader
* Groups now loaded as text by parquet-tools classes
* Added tests. Tests duplicated fot both implementation
