+++ src/test/resources/log4j.properties
@@ -1,0 +1,47 @@
+###
+# #%L
+# ICAT metadata ingestion core
+# %%
+# Copyright (C) 2014 ESRF - The European Synchrotron
+# %%
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+# 
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+# 
+# You should have received a copy of the GNU General Public
+# License along with this program.  If not, see
+# <http://www.gnu.org/licenses/gpl-3.0.html>.
+# #L%
+###
+#
+# The logging properties used
+#
+
+# log folder
+log.dir=log
+
+# Root logger option
+log4j.rootLogger=INFO, stdout, file
+
+# Framework logging
+
+# Client logging
+log4j.logger.tech.tablesaw=DEBUG
+ 
+# Direct log messages to a log file
+log4j.appender.file=org.apache.log4j.FileAppender
+log4j.appender.file.layout=org.apache.log4j.PatternLayout
+log4j.appender.file.layout.ConversionPattern=%d %-5p: [%t] %c{1} - %m%n
+
+### direct log messages to stdout ###
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{ABSOLUTE} %-5p: [%t] (%F:%L) - %m - (%c)%n
+
